// src/repository/openapi/model/category.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trust Score Dashboard API
 * 기업별 신뢰 점수 시계열 데이터 API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 카테고리
 * @export
 * @enum {string}
 */

export const Category = {
    RECRUIT: 'recruit',
    PERFORMANCE: 'performance',
    PUBLIC: 'public'
} as const;

export type Category = typeof Category[keyof typeof Category];