// src/repository/openapi/model/progress-stage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trust Score Dashboard API
 * 기업별 신뢰 점수 시계열 데이터 API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 계약 진행상태 (DB progress_stage 값과 일치)
 * @export
 * @enum {string}
 */

export const ProgressStage = {
    TEST: 'test',
    QUOTE: 'quote',
    APPROVAL: 'approval',
    CONTRACT: 'contract'
} as const;

export type ProgressStage = typeof ProgressStage[keyof typeof ProgressStage];